requirements.txt
# データベース
supabase>=2.0.0
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0
postgrest>=0.10.0

# データ処理
pandas>=2.0.0
numpy>=1.24.0
beautifulsoup4>=4.12.0
python-docx>=1.0.0
PyPDF2>=3.0.0

# 自然言語処理
mecab-python3>=1.0.6
spacy>=3.7.0
transformers>=4.36.0
langchain>=0.1.0

# ETL
apache-airflow>=2.7.0
dbt-postgres>=1.7.0

# 機械学習
scikit-learn>=1.3.0
statsmodels>=0.14.0

# API連携
notion-client>=2.0.0
openai>=1.0.0
grafana-api>=1.0.0

# ユーティリティ
python-dotenv>=1.0.0
pytest>=7.4.0
black>=23.0.0
flake8>=6.1.0
isort>=5.13.0
